Lesson8 modified code V1.1
