+++ src/main/java/com/potemkin/i/exception/MyNoSuchFieldException.java
@@ -1,0 +1,26 @@
+/*
+ * Все права защищены.
+ * Регулируется лицензией Epam.
+ * 
+ */
+package com.potemkin.i.exception;
+
+/**
+ * Класс исключения checked MyNoSuchFieldException, в случае отсутствии искомого
+ * поля (переменной) класса. Проверенные исключения необходимо объявить в
+ * предложении метода или конструктора, если они могут быть вызваны выполнением
+ * метода или конструктора и распространяться за пределы метода или
+ * конструктора.
+ * 
+ * @author Илья Пот
+ */
+public class MyNoSuchFieldException extends Exception {
+
+    /**
+     * Дефолтный конструктор класса. Создает новое исключение с со стандартным
+     * сообщением.
+     */
+    public MyNoSuchFieldException() {
+        super("Поле ненайдено (не существует)!");
+    }
+}
